Move files to the right dir
