+++ src/tc-renderer/store/modules/settings/default-settings.js
