+++ src/tc-renderer/store/modules/settings/validate-settings.js
